Support floating point and exponential notation coordinates

+++ src/main/kotlin/com/jzbrooks/avdo/graphic/command/HorizontalLineTo.kt
@@ -1,3 +1,3 @@
 package com.jzbrooks.avdo.graphic.command
 
-data class HorizontalLineTo(override val variant: VariantCommand.Variant, val arguments: List<Int>) : VariantCommand
+data class HorizontalLineTo(override val variant: VariantCommand.Variant, val arguments: List<Float>) : VariantCommand
