+++ src/main/kotlin/com/jzbrooks/avdo/graphic/command/VerticalLineTo.kt
@@ -1,3 +1,3 @@
 package com.jzbrooks.avdo.graphic.command
 
-data class VerticalLineTo(override val variant: VariantCommand.Variant, val arguments: List<Int>) : VariantCommand
+data class VerticalLineTo(override val variant: VariantCommand.Variant, val arguments: List<Float>) : VariantCommand
